Combine redditlink and userlink rules
