+++ src/rules/redditlink.ts
@@ -1,6 +1,7 @@
 import SimpleMarkdown from 'simple-markdown';
 import { SimpleMarkdownRule } from './ruleType.js';
 
+// Rule for both /?r/ and /?u/ links
 export const redditlink: SimpleMarkdownRule = {
 	order: SimpleMarkdown.defaultRules.url.order - 0.5,
 	match: function (source, state, prevCapture) {
@@ -8,7 +9,7 @@ export const redditlink: SimpleMarkdownRule = {
 			return null;
 		}
 		if (prevCapture.length !== 0 && prevCapture[prevCapture.length - 1] !== ' ') return null;
-		return source.match(/^\/?r\/\w{2,24}/);
+		return source.match(/^\/?r\/\w{2,24}|^\/?u\/\w{2,20}/);
 	},
 	parse: function (capture) {
 		let target = '';
